+++ CMakeLists.txt
@@ -54,27 +54,6 @@ set (Materials
 #
 
 add_definitions(-DFFX_CACAO_ENABLE_D3D12) # TODO: move render passes to renderer project
-set (RenderPassFiles
-    "Source/Engine/RenderPass/RenderPass.h"
-    "Source/Engine/RenderPass/AmbientOcclusion.h"
-    "Source/Engine/RenderPass/ApplyReflections.h"
-    "Source/Engine/RenderPass/DepthPrePass.h"
-    "Source/Engine/RenderPass/DepthMSAAResolve.h"
-    "Source/Engine/RenderPass/ScreenSpaceReflections.h"
-    "Source/Engine/RenderPass/MagnifierPass.h"
-    "Source/Engine/RenderPass/ObjectIDPass.h"
-    "Source/Engine/RenderPass/OutlinePass.h"
-    
-    "Source/Engine/RenderPass/RenderPass.cpp"
-    "Source/Engine/RenderPass/AmbientOcclusion.cpp"
-    "Source/Engine/RenderPass/ApplyReflections.cpp"
-    "Source/Engine/RenderPass/DepthPrePass.cpp"
-    "Source/Engine/RenderPass/DepthMSAAResolve.cpp"
-    "Source/Engine/RenderPass/ScreenSpaceReflections.cpp"
-    "Source/Engine/RenderPass/MagnifierPass.cpp"
-    "Source/Engine/RenderPass/ObjectIDPass.cpp"
-    "Source/Engine/RenderPass/OutlinePass.cpp"
-)
 
 set (CoreFiles 
     "Source/Engine/Core/Platform.h"
