+++ Source/Renderer/Core/Fence.cpp
@@ -28,10 +28,9 @@
 void Fence::Create(ID3D12Device* pDevice, const char* pDebugName)
 {
     mFenceValue = 0;
-    //ThrowIfFailed(
+    ThrowIfFailed(
         pDevice->CreateFence(0, D3D12_FENCE_FLAG_NONE, IID_PPV_ARGS(&ptr))
-    //)
-    ;
+    );
     SetName(ptr, pDebugName);
     //mHEvent = CreateEventEx(nullptr, "FENCE_EVENT0", FALSE, EVENT_ALL_ACCESS);
     mHEvent = CreateEventEx(nullptr, FALSE, FALSE, EVENT_ALL_ACCESS);
@@ -46,12 +45,9 @@ void Fence::Destroy()
 void Fence::Signal(ID3D12CommandQueue* pCommandQueue)
 {
     ++mFenceValue;
-    
-    //ThrowIfFailed(
+    ThrowIfFailed(
         pCommandQueue->Signal(ptr, mFenceValue)
-
-    //)
-    ;
+    );
 }
