+++ Source/Renderer/Rendering/RenderPass/AmbientOcclusion.cpp
