+++ Source/Renderer/Rendering/RenderPass/ApplyReflections.cpp
@@ -17,7 +17,7 @@
 //	Contact: volkanilbeyli@gmail.com
 
 #include "ApplyReflections.h"
-#include "../../Renderer/Renderer.h"
+#include "Renderer/Renderer.h"
 
 ApplyReflectionsPass::ApplyReflectionsPass(VQRenderer& Renderer)
 	: RenderPassBase(Renderer)
