+++ Source/Renderer/Rendering/RenderPass/MagnifierPass.h
